requirements.txt
pydantic==2.3.0
more_itertools==10.1.0
numpy~=1.26.2
pandas==2.1.1
SimpleITK==2.3.0
black==23.9.1
scipy==1.11.2
pydicom==2.4.3
opencv-python>=4.8
matplotlib>=3.8
PyQt5==5.15.10
PyYAML~=6.0.1
jupyter==1.1.1
mplcursors~=0.5.2
ipympl==0.9.4
shapely==2.0.2
thefuzz==0.22.1
Levenshtein==0.26.1
torch==2.5.1
tensorboard==2.18.0
torchvision==0.20.1
seaborn==0.13.2
tqdm==4.67.0
click==8.1.7
scikit-learn==1.5.2